test: move regression tests to test/regression/issue/10411.test.ts
The gate runs the whole test file and run-autoinstall.test.ts has pre-existing
'basic autoinstall' tests that fetch is-even from npmjs.org and flake at the
5s default under ASAN. The new tests are hermetic (loopback registry, isolated
cache dir), so a dedicated file avoids unrelated timeouts.
